Removed need for separate GS Storage Bean Configuration
