+++ gameserver/src/main/java/dev/totallyspies/spydle/gameserver/storage/CurrentGameServerConfiguration.java
@@ -4,11 +4,15 @@ import dev.totallyspies.spydle.gameserver.agones.AgonesHook;
 import dev.totallyspies.spydle.shared.model.GameServer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.boot.context.event.ApplicationReadyEvent;
+import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
+import org.springframework.context.event.EventListener;
 
 /**
  * Configuration to decide which bean should declare how we get information about our current gameserver.
@@ -20,8 +24,14 @@ public class CurrentGameServerConfiguration {
 
     private final Logger logger = LoggerFactory.getLogger(CurrentGameServerConfiguration.class);
 
+    @Autowired
+    private ApplicationContext applicationContext;
+
+    @Autowired
+    private GameServerStorage storage;
+
     @Bean
-    @Primary // Fallback option
+    @Primary
     @ConditionalOnProperty(name = "agones.enabled", havingValue = "true")
     public GameServer currentAgonesGameServer(AgonesHook agonesHook) {
         logger.info("Agones has been enabled in application properties, loading agones current game server info...");
@@ -40,6 +50,13 @@ public class CurrentGameServerConfiguration {
                 .publicRoom(false) // TODO
                 .state(GameServer.State.WAITING)
                 .build();
+    }
+
+    @EventListener(ApplicationReadyEvent.class)
+    public void onApplicationReady() {
+        GameServer currentGameServer = applicationContext.getBean(GameServer.class);
+        storage.storeGameServer(currentGameServer);
+        logger.info("Wrote current game server to storage: {}", currentGameServer);
     }
 
 }
